requirements.txt
asgiref==3.5.2
attrs==22.1.0
colorama==0.4.6
Django==2.2.16
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.2
importlib-metadata==5.0.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
zipp==3.10.0